spans, annotated formattedref: https://github.com/metanorma/metanorma-iso/issues/681
