+++ lib/metanorma/standoc/cleanup.rb
@@ -5,7 +5,7 @@ require_relative "./cleanup_block"
 require_relative "./cleanup_table"
 require_relative "./cleanup_footnotes"
 require_relative "./cleanup_ref"
-require_relative "./cleanup_ref_dl"
+require_relative "./cleanup_biblio"
 require_relative "./cleanup_boilerplate"
 require_relative "./cleanup_section"
 require_relative "./cleanup_terms"
